requirements.txt
opencv-python==4.7.0.72
pillow==8.4.0
tensorflow==2.10.0
openai==1.12.0